rename file and updates
